Rename variables
I feel that the variable names are a little too open, and worry that they might conflict in my existing application either now or in the future. I'd like to proprose renaming the variables used within this plugin to avoid clashes.

I realise this is a breaking change, so would need to be included in a new major release.